dom/media/test/test_mediarecorder_creation_fail.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test MediaRecorder Record with media.ogg.enabled = false</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <script type="text/javascript" src="manifest.js"></script>
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">
function testThrows(stream, options) {
  try {
    new MediaRecorder(stream, options);
    return false;
  } catch {
    return true;
  }
}
(async () => {
  SimpleTest.waitForExplicitFinish();
  await SpecialPowers.pushPrefEnv({set: [
      ["media.ogg.enabled", false],
      ["media.encoder.webm.enabled", false],
    ]});
  const stream = new AudioContext().createMediaStreamDestination().stream;
  ok(testThrows(stream, {mimeType: "audio/ogg"}),
    "Creating an ogg recorder without ogg support throws");
  ok(testThrows(stream, {mimeType: "audio/webm"}),
    "Creating a webm recorder without webm support throws");
  ok(testThrows(stream, {mimeType: "video/webm"}),
    "Creating a webm recorder without webm support throws");
  ok(testThrows(stream, {mimeType: "apa/bepa"}),
    "Creating a recorder for a bogus mime type throws");
  ok(testThrows(stream, {}),
    "Creating a default recorder without any container support throws");
  SimpleTest.finish();
})();
</script>
</pre>
</body>
</html>
